+++ service/database/migrations/7_add-token-redeemed-at.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `token`
+DROP COLUMN `redeemed_at`;